requirements.txt
autobahn
twisted
alpha_mini_rug
google.genai
pyopenssl
python-dotenv
requests